+++ sdk/src/main/java/ly/count/android/sdk/ConnectionQueue.java
@@ -31,6 +31,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import org.json.JSONException;
@@ -52,6 +53,9 @@ class ConnectionQueue implements RequestQueueProvider {
     private Future<?> connectionProcessorFuture_;
     private DeviceIdProvider deviceIdProvider_;
     private SSLContext sslContext_;
+    private final ScheduledExecutorService backoffScheduler_ = Executors.newSingleThreadScheduledExecutor();
+    private final AtomicBoolean backoff_ = new AtomicBoolean(false);
+
     BaseInfoProvider baseInfoProvider;
     HealthTracker healthTracker;
     public PerformanceCounterCollector pcc;
@@ -874,6 +878,10 @@ class ConnectionQueue implements RequestQueueProvider {
     public void tick() {
         //todo enable later
         //assert storageProvider != null;
+        if (backoff_.get()) {
+            L.i("[ConnectionQueue] tick, currently backed off, skipping tick");
+            return;
+        }
 
         boolean rqEmpty = isRequestQueueEmpty(); // this is a heavy operation, do it only once. Why heavy? reading storage
         boolean cpDoneIfOngoing = connectionProcessorFuture_ != null && connectionProcessorFuture_.isDone();
@@ -893,7 +901,21 @@ class ConnectionQueue implements RequestQueueProvider {
     }
 
     public ConnectionProcessor createConnectionProcessor() {
-        ConnectionProcessor cp = new ConnectionProcessor(baseInfoProvider.getServerURL(), storageProvider, deviceIdProvider_, configProvider, requestInfoProvider, sslContext_, requestHeaderCustomValues, L, healthTracker);
+
+        ConnectionProcessor cp = new ConnectionProcessor(baseInfoProvider.getServerURL(), storageProvider, deviceIdProvider_, configProvider, requestInfoProvider, sslContext_, requestHeaderCustomValues, L, healthTracker, new Runnable() {
+            @Override
+            public void run() {
+                L.d("[ConnectionQueue] createConnectionProcessor:run, backed off, countdown started for " + configProvider.getBOMDuration() + " seconds");
+                backoff_.set(true);
+                backoffScheduler_.schedule(new Runnable() {
+                    @Override public void run() {
+                        L.d("[ConnectionQueue] createConnectionProcessor:run, countdown finished, running tick in background thread");
+                        backoff_.set(false);
+                        tick();
+                    }
+                }, configProvider.getBOMDuration(), TimeUnit.SECONDS);
+            }
+        });
         cp.pcc = pcc;
         return cp;
     }